task_service/config.py
import os
from dotenv import load_dotenv

load_dotenv()

DB_USER = os.getenv("DB_USER")
DB_PASSWORD = os.getenv("DB_PASSWORD")
DB_NAME = os.getenv("DB_NAME")
DB_HOST = os.getenv("DB_HOST")
DB_PORT = os.getenv("DB_PORT")

print("DB_HOST:", os.getenv("DB_HOST"))
SQLALCHEMY_DATABASE_URI = f"postgresql+psycopg2://{DB_USER}:{DB_PASSWORD}@{DB_HOST}:{DB_PORT}/{DB_NAME}"
print("DB_HOST:", os.getenv("DB_HOST"))

JWT_SECRET_KEY = os.getenv("JWT_SECRET_KEY")
CORS_ALLOWED_ORIGIN = os.getenv("CORS_ALLOWED_ORIGIN")